requirements.txt
ultralytics>=8.1.0
opencv-python>=4.8.0
pandas>=2.0.0
requests>=2.31.0
pillow>=10.0.0
matplotlib>=3.7.0
ipython>=8.12.0
